[IMP] hr_holidays: improve display name
Before this commit, the date range was not known without open the record in backend.
It is really painfull to don't know when a guy of your team ping you on a new leave
the date from the current leave. (today ? next week ? ...)

Now we have the date into the name_get and so the subject of the email.

task-2001439

Nothing related to next activity ;) but idea was the same.

closes odoo/odoo#44462

Signed-off-by: Jérémy Kersten (jke) <jke@openerp.com>