[FIX] sale: fix traceback when creating an INV for an empty SO
Currently when creating an SO without any products and creating
down payment for that creates a traceback.

so Fix traceback when creating an invoice for sale order with an empty
product. when there is no orderline then do not get the sequence in down
payment line to make down payment line at last in sales order.

Task-2181464

closes odoo/odoo#44645

Closes: #44491
X-original-commit: d839d2023bdec578ed96a8358623125165d67d39
Signed-off-by: Yannick Tivisse (yti) <yti@odoo.com>