[IMP] fleet: Close driver history on driver change
Purpose
=======
If you set a future driver and click on "apply change", a new assignation log is
automatically created for the new driver.
But the end date of the former assignation isn't populate.

Specifications
==============
Set the date "end date" for the former driver with the "start date" of the
new driver.

closes odoo/odoo#44635

Signed-off-by: Yannick Tivisse (yti) <yti@odoo.com>