[FIX] website_form, *: fix new website form
* = website_crm, website_hr_recruitment, website_sale

Changing label position on existing field was replacing the label with
the default one.

Redirection message was not updating the overlay

Going from show message to redirect without hidding the mesage made the
form d-none

When a select field's label's position was changed the select was not
editable anymore.

Some other miscellaneous fixes

task-2191409

closes odoo/odoo#44911

X-original-commit: 86b5a3585d1e68c4516bcc0a258ec0f01b211e37
Signed-off-by: Quentin Smetz (qsm) <qsm@odoo.com>