[FIX] website_event_sale: reset sale when buying free tickets
When we buy a ticket on website, a sale order is automatically created and
confirmed if the ticket is free. When we buy a second ticket we will edit the
old SO, but we can't, because the SO was confirmed.

Solution
========

In `_create_attendees_from_registration_post`, reset the sale information.

Task ID 2188857
PR #44545

X-original-commit: c30299a9e70c20054e2e43f3e727a4d48e66a571